comments for multi send flow
